Finished up the migration to expressions
Expressions are defined as operations that return a variable where commands are operations that change the program state. This distinction will greatly help moving forward.

all mathematical operations now work correctly and we now support expression nesting as in x = (y+z) + 3

A side affect of this is that we can now declare scopes and functions much easier

Next large TODO is the implementation of functions