Comparators, better if statemens, while loops
I added the less than, greater than, and equality comparators. They
are expressions that return bool values based on the values of what
is passed in.

I changed if statements to take expressions and be more in line with
the code rework that differentiated between expressions and commands

I added while loops as syntactic sugar and set up the framwork for for
loops

I added a few examples and made the examples a little more readable