Refactor cmd, expr
Refactor expr and cmd, creating the CmdIter and ExprCmp data types for
commands that iterate variables and expressions that compare variables
respectively. This allows us to reduce undefined behavior in for and
while loops by forcing the type of expressions that the user must use.

Reintroduce the comparison expressions and the iteration commands

TODO is loops and functions follows by static type checking with
the new BinOps