Framework for type checking, syntactic sugar
Further set up the framework for type checking - each program is checked at compile time for type errors. The overall program can either return the result state (if the type checking was successful) or an error message representing what was wrong with the program's typing.

the typeCheck function checks commands and returns a maybe string - an error string if the program is invalid and nothing if the program is okay

the typeExpr function either returns the final type of an expression or nothing if the expression has an error at some point

For ease of type checking as well as simplicity in design, for loops and increment operators have been temporarily removed to be added as syntactic sugar in the future.

TODO: Fill out type checking
TODO: Reinstate for loops and increment operators